Later implementations (back end)

+++ src/routes/stockRoutes.ts
@@ -1,17 +1,27 @@
 import express, { Request, Response } from 'express';
+import rateLimit from 'express-rate-limit';
 import TypeLogs from "../log/TypeLogs";
 import StockController from '../controller/StockController';
+import { comandsTables as tables } from '../model/StockTables'; // Importando as tabelas configuradas
 
 const router = express.Router();
 const stockController = new StockController(new TypeLogs());
 let message: string;
 
+function isValidTable(nameTable: string): nameTable is keyof typeof tables {
+    return nameTable in tables;
+}
+
 router.get('/all/:nameTable', async (req: Request, res: Response) => {
     const { nameTable } = req.params;
     let userIP = req.ip === "::1" ? "127.0.0.1" : req.ip;
-    message = "All itens retrieved";
+    message = "All items retrieved";
 
     try {
+        if (!isValidTable(nameTable)) {
+            throw new Error(`Invalid table name: ${nameTable}`);
+        }
+
         const result = await stockController.allItens(nameTable);
 
         stockController.getLogger().info(message, userIP);
@@ -23,15 +33,17 @@ router.get('/all/:nameTable', async (req: Request, res: Response) => {
     }
 });
 
-
-router.post('/add/:nameTable', (req: Request, res: Response) => {
+router.post('/add/:nameTable', async (req: Request, res: Response) => {
     const { nameTable } = req.params;
     let userIP = req.ip === "::1" ? "127.0.0.1" : req.ip;
     message = "Add successful";
 
     try {
-        const data = stockController;
+        if (!isValidTable(nameTable)) {
+            throw new Error(`Invalid table name: ${nameTable}`);
+        }
 
+        await stockController.addItem(req.body, nameTable);
 
         stockController.getLogger().info(message, userIP);
 
@@ -42,16 +54,17 @@ router.post('/add/:nameTable', (req: Request, res: Response) => {
     }
 });
 
-
 router.put('/update/:nameTable/:idProduct', async (req: Request, res: Response) => {
     const { nameTable, idProduct } = req.params;
     let userIP = req.ip === "::1" ? "127.0.0.1" : req.ip;
     message = "Update successful";
 
     try {
-        const data = stockController.validateTableData(nameTable, idProduct, req.body);
+        if (!isValidTable(nameTable)) {
+            throw new Error(`Invalid table name: ${nameTable}`);
+        }
 
-        await stockController.update(data, nameTable);
+        await stockController.updateItem(req.body, nameTable, parseInt(idProduct, 10));
 
         stockController.getLogger().warn(message, userIP);
 
@@ -62,18 +75,42 @@ router.put('/update/:nameTable/:idProduct', async (req: Request, res: Response) 
     }
 });
 
-
 router.delete('/delete/:nameTable/:idProduct', async (req: Request, res: Response) => {
     const { nameTable, idProduct } = req.params;
     let userIP = req.ip === "::1" ? "127.0.0.1" : req.ip;
     message = "Delete successful";
 
     try {
+        if (!isValidTable(nameTable)) {
+            throw new Error(`Invalid table name: ${nameTable}`);
+        }
+
         await stockController.deleteItem(nameTable, parseInt(idProduct, 10));
-        
+
         stockController.getLogger().warn(message, userIP);
-        
+
         res.status(200).send({ messageStatus: message });
+    } catch (error) {
+        stockController.getLogger().error((error as Error).message, userIP);
+        res.status(500).json({ error: (error as Error).message });
+    }
+});
+
+router.put('/update-quantity-total/:nameTable/:idProduct/:typeMoviment' , async (req: Request, res: Response) => {
+    const { nameTable, idProduct, typeMoviment } = req.params;
+    let userIP = req.ip === "::1" ? "127.0.0.1" : req.ip;
+    message = "Total balance update successful";
+
+    try {
+        if (!isValidTable(nameTable)) {
+            throw new Error(`Invalid table name: ${nameTable}`);
+        }
+
+        await stockController.updateQuantityTotal(parseInt(req.body["quantity"], 10), nameTable, parseInt(idProduct, 10), typeMoviment);
+
+        stockController.getLogger().warn(message, userIP);
+
+        res.status(201).json({ messageStatus: message });
     } catch (error) {
         stockController.getLogger().error((error as Error).message, userIP);
         res.status(500).json({ error: (error as Error).message });
